requirements.txt
PyQt5==5.14.1
pillow==10.4.0
requests==2.27.1
